Modified project structure. New files added.
