+++ GOOGLE EXERCISES/Course 1/Mod4_Strings-1.py
