+++ GOOGLE EXERCISES/Course 1/Mod4_Strings-2.py
