+++ GOOGLE EXERCISES/Course 2/Mod2_aaaaaa.py
@@ -1,0 +1,9 @@
+import shutil
+du = shutil.disk_usage("/")
+print(du)
+percentage = du.free/du.total*100
+print(percentage)
+# ==============================
+import psutil
+cpuPerc = psutil.cpu_percent(0.1)
+print(cpuPerc)
