+++ GOOGLE EXERCISES/Course 2/Mod2_cccc.py
@@ -1,0 +1,5 @@
+name = "Bob"
+position = "worker"
+phone = "123"
+
+print("Name: {}, position: {}, Phone: {}".format(name, position, phone))
